requirements.txt
numpy>=1.20
Pillow>=8.0
matplotlib>=3.4
tqdm>=4.60
